-- Feign Client integration -- Unit Test implementation -- Integration Tests for Customer Services -- Integration Tests for Customer Controller -- Validation for Customer

+++ src/main/java/com/credit/app/core/persist/Customer.java
@@ -1,9 +1,11 @@
 package com.credit.app.core.persist;
 
+import com.credit.app.core.common.constants.Messages;
 import com.credit.app.core.common.model.CustomerInfo;
 import lombok.Data;
 
 import javax.persistence.Entity;
+import javax.validation.constraints.NotEmpty;
 
 
 @Data
@@ -13,9 +15,9 @@ public class Customer extends BaseEntity<CustomerInfo> {
     private String email;
     private String name;
     private String surname;
-    private long salary;
+    private double salary;
     private String telephone;
-    private long creditLimit;
+    private double creditLimit;
 
     @Override
     public CustomerInfo toInfo() {
